Modified install script
When extracting the installers from the wrapper, the 32 and 64 bit installers for the extension are renamed to distinguish them. The 64 bit installer then installs the 64 bit extension and falls through to the 32 bit installer, which installs the proxy and the 32 bit extension.

Fixes #5 : Install 32 and 64 bit support on 64 bit systems