Save As dialog no longer asks to overwrite new files - google/google-drive-shell-extension#12
Added three macros to check constraints and homogenize error signaling.

Removed a block that was adding newly created files to the ignoredIds map and preventing the explorer view from showing the new file.

Corrected an error return in _GetDataFromID from invalid argument to file not found.

Simplified the code for GetAttributesOf.