Possible fix to crash on Save As
We were freeing the value of *ppidl before assigning to it. This has no effect when *ppidl was nullptr, and simply freed the memory if it was a pointer to a valid object. However, being an output parameter, there is no promise that *ppidl will not contain garbage. It is possible that this was causing the reported crash. This has been fixed.

Improved reporting in case the issue lays elsewhere.