Fix for issue where stub files are created while browsing.
When browsing from the Save As dialog to a save directory within the extension, empty files were being created along the path in each directory. The reason seemed to be that ParseDisplayName was called with the STGM_CREATE mode option, which is supposed to return a valid pidl but not modify the store.

This commit pushes the creation of the file back to when the shell tries to check the properties of the file, in the call to GetAttributesOf.